Fix equals
Now assigns off-display operand the same value as on-display operand if
the former has no value. This is only the case in the specific sequence
of entering a number, choosing an operand, and pressing equals. In
practice this means that the mentioned sequence now leads to the
behaviour of (a `op` a) with immediate responsiveness.