Fix bug & simplify logic
operandOffDisplay-assignment added to changeOperator() and removed from
addDigit(). Once an operator is pressed the current display value is the
de facto operand to-be stored off-display. Moving this from addDigit()
to changeOperator() makes it so that a sequence
"a `op` b `another_op` `equals`" now results in "b `another_op` b" after
the equals, which prior to this change would result in
"b `another_op` a" once before repeatedly turning into
"b `another_op` b" with subsequenct equals.